move tests to appropriate places
