+++ containers/ecr-viewer/src/app/tests/services/evaluateFhirDataServices.ts
@@ -1,0 +1,84 @@
+import { loadYamlConfig } from "@/app/api/utils";
+import {
+  evaluateReference,
+  evaluateValue,
+} from "@/app/services/evaluateFhirDataService";
+import BundleWithMiscNotes from "@/app/tests/assets/BundleMiscNotes.json";
+import { Bundle } from "fhir/r4";
+import BundleWithPatient from "@/app/tests/assets/BundlePatient.json";
+
+const mappings = loadYamlConfig();
+
+describe("Evaluate Reference", () => {
+  it("should return undefined if resource not found", () => {
+    const actual = evaluateReference(
+      BundleWithMiscNotes as unknown as Bundle,
+      mappings,
+      "Observation/1234",
+    );
+
+    expect(actual).toBeUndefined();
+  });
+  it("should return the resource if the resource is available", () => {
+    const actual = evaluateReference(
+      BundleWithPatient as unknown as Bundle,
+      mappings,
+      "Patient/6b6b3c4c-4884-4a96-b6ab-c46406839cea",
+    );
+
+    expect(actual.id).toEqual("6b6b3c4c-4884-4a96-b6ab-c46406839cea");
+    expect(actual.resourceType).toEqual("Patient");
+  });
+});
+
+describe("evaluate value", () => {
+  it("should provide the string in the case of valueString", () => {
+    const actual = evaluateValue(
+      { resourceType: "Observation", valueString: "abc" } as any,
+      "value",
+    );
+
+    expect(actual).toEqual("abc");
+  });
+  it("should provide the string in the case of valueCodeableConcept", () => {
+    const actual = evaluateValue(
+      {
+        resourceType: "Observation",
+        valueCodeableConcept: {
+          coding: [
+            {
+              display: "Negative",
+            },
+          ],
+        },
+      } as any,
+      "value",
+    );
+
+    expect(actual).toEqual("Negative");
+  });
+  describe("Quantity", () => {
+    it("should provide the value and string unit with a space inbetween", () => {
+      const actual = evaluateValue(
+        {
+          resourceType: "Observation",
+          valueQuantity: { value: 1, unit: "ft" },
+        } as any,
+        "value",
+      );
+
+      expect(actual).toEqual("1 ft");
+    });
+    it("should provide the value and symbol unit", () => {
+      const actual = evaluateValue(
+        {
+          resourceType: "Observation",
+          valueQuantity: { value: 1, unit: "%" },
+        } as any,
+        "value",
+      );
+
+      expect(actual).toEqual("1%");
+    });
+  });
+});
